AnnotationEditor plugin initialized
Copied over relevant files from Afra, while commenting out or slightly
modifying part of the code (especially in EditTrack and ScratchPad)
that don't work (such as syncing to localStorage and validations
that are run after edit operations).

Afra's CSS is a mess. Styles relevant for the plugin are hard to
identify all at once. So only a subset of all styles has been
copied over as a starting point. For jquery-ui, we downloaded
a custom build that only contains the CSS and JS for drggable,
droppable and resizable and used that here.

Currently, we are able to drag and drop feature to edit track
and resize exons in edit track (this errors out at the moment).

Signed-off-by: aniarya82 <hiten.iitkgp@gmail.com>
Signed-off-by: Anurag Priyam <anurag.priyam@qmul.ac.uk>